Some movements and renaming ts
