+++ src/VSCode.Extension/extension.ts
@@ -1,12 +1,12 @@
-import { DotNetDebuggerConfiguration } from './tasks/debug';
-import { ConfigurationController } from './configuration';
-import { DotNetTaskProvider } from './tasks/build';
-import { XamlController } from './xaml/service';
-import { CommandController } from './bridge';
-import { UIController } from './controller';
-import { PublicExports } from './exports';
-import { StateController } from './cache';
-import * as res from './resources';
+import { MonoDebugConfigurationProvider } from './tasks/monoDebugConfigurationProvider';
+import { DotNetTaskProvider } from './tasks/dotnetTaskProvider';
+import { ConfigurationController } from './configurationController';
+import { StatusBarController } from './statusbarController';
+import { CommandController } from './commandController';
+import { StateController } from './stateController';
+import { XamlController } from './xaml/xamlController';
+import { PublicExports } from './publicExports';
+import * as res from './resources/constants';
 import * as vscode from 'vscode';
 
 
@@ -20,30 +20,30 @@ export function activate(context: vscode.ExtensionContext): PublicExports | unde
 	ConfigurationController.activate(context);
 	StateController.activate(context);
 	XamlController.activate(context);
-	UIController.activate(context);
-	UIController.update();
+	StatusBarController.activate(context);
+	StatusBarController.update();
 
 	/* Commands */
-	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdSelectActiveProject, UIController.showQuickPickProject));
-	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdSelectActiveConfiguration, UIController.showQuickPickTarget));
-	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdSelectActiveDevice, UIController.showQuickPickDevice));
+	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdSelectActiveProject, StatusBarController.showQuickPickProject));
+	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdSelectActiveConfiguration, StatusBarController.showQuickPickTarget));
+	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdSelectActiveDevice, StatusBarController.showQuickPickDevice));
 	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdActiveTargetFramework, () => ConfigurationController.getTargetFramework()));
 	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdActiveConfiguration, () => ConfigurationController.target));
 	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdActiveProjectPath, () => ConfigurationController.project?.path));
 	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdActiveDeviceName, () => ConfigurationController.device?.name));
 	context.subscriptions.push(vscode.commands.registerCommand(res.commandIdActiveDeviceSerial, () => ConfigurationController.device?.serial));
 	/* Providers */
-	context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider(res.debuggerMeteorId, new DotNetDebuggerConfiguration()));
+	context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider(res.debuggerMeteorId, new MonoDebugConfigurationProvider()));
 	context.subscriptions.push(vscode.tasks.registerTaskProvider(res.taskDefinitionId, new DotNetTaskProvider()));
 	/* Events */
 	context.subscriptions.push(vscode.workspace.onDidSaveTextDocument(ev => {
 		if (ev.fileName.endsWith('.xaml') && vscode.debug.activeDebugSession?.configuration.type === res.debuggerMeteorId)
 			CommandController.xamlReload(ConfigurationController.getReloadHostPort(), ev.fileName);
 	}));
-	context.subscriptions.push(vscode.workspace.onDidChangeWorkspaceFolders(UIController.update));
+	context.subscriptions.push(vscode.workspace.onDidChangeWorkspaceFolders(StatusBarController.update));
 	context.subscriptions.push(vscode.workspace.onDidSaveTextDocument(ev => {
 		if (ev.fileName.endsWith('proj') || ev.fileName.endsWith('.props'))
-			UIController.update();
+			StatusBarController.update();
 	}));
 
 	return exports;
