+++ src/VSCode.Extension/models/project.ts
@@ -1,0 +1,7 @@
+
+export interface Project {
+    name: string;
+    path: string;
+    frameworks: string[];
+}
+
