+++ src/VSCode.Extension/models/target.ts
@@ -1,0 +1,5 @@
+
+export enum Target {
+    Debug = "Debug",
+    Release = "Release"
+}
