+++ src/VSCode.Extension/resources/constants.ts
