+++ src/VSCode.Extension/stateController.ts
@@ -1,12 +1,11 @@
-import { ConfigurationController } from './configuration';
-import { UIController } from './controller';
+import { ConfigurationController } from './configurationController';
+import { StatusBarController } from './statusbarController';
 import { ExtensionContext } from 'vscode';
-import * as models from "./models"
-
+import { Target } from './models/target';
+import { Device } from './models/device';
 
 export class StateController {
     private static context: ExtensionContext | undefined;
-
 
     public static activate(context: ExtensionContext) {
         StateController.context = context;
@@ -15,17 +14,16 @@ export class StateController {
         StateController.context = undefined;
     }
 
-
     public static load() {
         if (StateController.context === undefined)
             return;
 
         const project = StateController.context.workspaceState.get<string>('project');
         const device = StateController.context.workspaceState.get<string>('device');
-        const target = StateController.context.workspaceState.get<models.Target>('target');
+        const target = StateController.context.workspaceState.get<Target>('target');
 
-        ConfigurationController.device = UIController.devices.find(it => StateController.getDeviceId(it) === device);
-        ConfigurationController.project = UIController.projects.find(it => it.path === project);
+        ConfigurationController.device = StatusBarController.devices.find(it => StateController.getDeviceId(it) === device);
+        ConfigurationController.project = StatusBarController.projects.find(it => it.path === project);
         ConfigurationController.target = target;
     }
     public static saveProject() {
@@ -41,7 +39,7 @@ export class StateController {
             StateController.context.workspaceState.update('target', ConfigurationController.target);
     }
 
-    private static getDeviceId(device: models.IDevice | undefined): string {
+    private static getDeviceId(device: Device | undefined): string {
         return device ? `${device.name}_${device.platform}_${device.os_version}` : 'null';
     }
 }
