+++ src/VSCode.Extension/xaml/models/context.ts
@@ -1,0 +1,7 @@
+
+export class Context {
+    public name: string | undefined;
+    public prefix: string | undefined;
+    public parent: Context | undefined;
+    public namespace: string | undefined;
+}
