+++ src/VSCode.Extension/xaml/models/xamlContext.ts
@@ -1,0 +1,9 @@
+import { Context } from "./context";
+import { XamlScope } from "./xamlScope";
+
+export class XamlContext {
+    public tagContext: Context | undefined;
+    public attributeContext: Context | undefined;
+    public imports: { [prefix: string]: string } = {};
+    public scope: XamlScope | undefined;
+}
