+++ src/VSCode.Extension/xaml/models/xamlScope.ts
@@ -1,0 +1,8 @@
+
+export enum XamlScope {
+    Tag = 0,
+    Attribute = 1,
+    Multiline = 2,
+    Static = 3,
+    Value = 4
+}
