+++ src/VSCode.Extension/xaml/xamlCommandController.ts
@@ -1,14 +1,15 @@
-import { ProcessArgumentBuilder, ProcessRunner } from "../bridge";
-import * as res from '../resources';
+import { ProcessArgumentBuilder } from '../processes/processArgumentBuilder';
+import { ProcessRunner } from '../processes/processRunner';
+import * as res from '../resources/constants';
 import * as vscode from 'vscode';
 import * as path from "path";
 import * as fs from "fs";
 
-export class CommandController {
+export class XamlCommandController {
     private static xamlToolPath: string;
 
     public static activate(context: vscode.ExtensionContext) {
-        const qualifiedVersion = CommandController.runtimeVersion() ?? "";
+        const qualifiedVersion = XamlCommandController.runtimeVersion() ?? "";
         const qualifiedVersionRegex = new RegExp('^\\d+\\.\\d+', ''); 
         const versionRegexCollection = qualifiedVersionRegex.exec(qualifiedVersion);
         const version = (versionRegexCollection && versionRegexCollection.length !== 0) 
@@ -20,11 +21,11 @@ export class CommandController {
             extensionBinaryPath = path.join(extensionPath, "extension", "bin", "Xaml", "net6.0");
         
         const executableExtension = process.platform === 'win32' ? '.exe' : '';
-        CommandController.xamlToolPath = path.join(extensionBinaryPath, "DotNet.Meteor.Xaml" + executableExtension);
+        XamlCommandController.xamlToolPath = path.join(extensionBinaryPath, "DotNet.Meteor.Xaml" + executableExtension);
     }
 
     public static async xamlSchema(path: string): Promise<boolean>  {
-        return await ProcessRunner.runAsync<boolean>(new ProcessArgumentBuilder(CommandController.xamlToolPath)
+        return await ProcessRunner.runAsync<boolean>(new ProcessArgumentBuilder(XamlCommandController.xamlToolPath)
             .appendQuoted(path));
     }
     private static runtimeVersion(): string | undefined {
